fix twice reading data from web-worker
