+++ react/react-ts/src/features/notes/FilterNotes/model/filterNotes.ts
@@ -1,21 +1,11 @@
 import {createAsyncThunk} from "@reduxjs/toolkit";
 import {RootState} from "@/app/appStore";
-import {FilterDataEvent, Note} from "@/entities/NoteList/model/types";
-import {noteListActions} from "@/entities/NoteList/model/slice";
+import {FilterDataEvent} from "@/entities/NoteList/model/types";
 import {worker} from "@/entities/NoteList/model/webWorketInit";
 
-export const filterNotes = createAsyncThunk<void, Note[], { state: RootState }>(
+export const filterNotes = createAsyncThunk<void, FilterDataEvent, { state: RootState }>(
     'notes/filterNotes',
-    async (notes, {dispatch}) => {
-        const {updateNoteList} = noteListActions;
-        const filterDataEvent: FilterDataEvent = {
-            cmd: "filterData",
-            payload: "blue"
-        }
-        worker.postMessage(filterDataEvent)
-        worker.onmessage = (e) => {
-            const sortedNotes: Note[] = e.data;
-            dispatch(updateNoteList(sortedNotes))
-        }
+    async (event, {dispatch}) => {
+        worker.postMessage(event)
     }
 )
